+++ tests/impls/yolo/test_yolo_results.py
@@ -60,8 +60,8 @@ def yolo_results(version: int) -> None:
 
     engine_path = build_yolo(version)
 
-    scale = True if version != 0 else False
-    trt_model = trtutils.impls.yolo.YOLO(engine_path, warmup=False, scale_inputs=scale)
+    scale = (0, 1) if version != 0 else (0, 255)
+    trt_model = trtutils.impls.yolo.YOLO(engine_path, warmup=False, input_range=scale)
 
     for gt, ipath in zip(
         [1, 3],
@@ -95,6 +95,11 @@ def test_yolo_9_results():
 
 def test_yolo_10_results():
     yolo_results(10)
+
+
+# cannot build yolox engine with trtexec
+# def test_yolo_x_results():
+#     yolo_results(0)
 
 
 if __name__ == "__main__":